refactor: cli 도구에서 문자열 치환 관련 명령어의 액션 함수에서 공통 동작 부분을 별도의 함수로 리팩토링
